// generated/src/app/api-models/al-letter-request-setup.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class AlLetterRequestSetup {
  seqLetterRequestId: number;
  jobId: string;
  requestUser: string;
  requestDate: Date;
  exportDirectory: string;
  exportFilePrefix: string;
  letterType: string;
  letterEventId: string;
  ruleOrder: number;
  memberSelection: number;
  fromDate: Date;
  thruDate: Date;
  action: number;
  status: number;
  requestType: string;
  requestStartTime: Date;
  requestTranInterval: number;
  comments: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}